backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
langgraph>=0.2.0,<1.0.0
langchain-google-genai>=2.1.0,<3.0.0
langchain-core>=0.3.0,<1.0.0
langsmith>=0.2.0,<1.0.0
python-frontmatter>=1.1.0,<2.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
numpy>=1.26.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
pypdf>=5.1.0,<6.0.0
python-docx>=1.1.0,<2.0.0
